+++ RELEASE_NOTES.md
@@ -16,3 +16,4 @@ test base
 * EspressoTestCase - base for default espresso tests
 * EspressoIntentTestCase - base test case when for mocking intents
 * CloseAllActivitiesFunction - use it to ensure that all started activities are finished
+* EspDummyLauncherActivity - support back press at start activity